requirements.txt
# Pin after `pip install -r requirements.txt && pip freeze > requirements.lock.txt` on your machine.
ultralytics>=8.3.0,<9
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
scikit-learn>=1.3.0
tqdm>=4.65.0
